"simplified all hit and readout classes and their tests; updates #107"
