--- prototype2/gdgem/nmx/HitFileTest.cpp
@@ -1,28 +1,0 @@
-/** Copyright (C) 2016, 2017 European Spallation Source ERIC */
-
-#include <gdgem/nmx/HitFile.h>
-#include <test/TestBase.h>
-
-class NMXHitFileTest : public TestBase {
-protected:
-  virtual void SetUp() {
-    hdf5::error::Singleton::instance().auto_print(false);
-    if (boost::filesystem::exists("hit_file_test.h5"))
-    {
-      boost::filesystem::remove("hit_file_test.h5");
-    }
-  }
-  virtual void TearDown() {}
-};
-
-TEST_F(NMXHitFileTest, CreateFile) {
-  HitFile::create("hit_file_test");
-  EXPECT_TRUE(hdf5::file::is_hdf5_file("hit_file_test.h5"));
-}
-
-// \todo test h5 compound type mapping
-
-int main(int argc, char **argv) {
-  testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
