+++ prototype2/multigrid/CMakeLists.txt
@@ -35,7 +35,6 @@ set(mgmesytec_common_inc
   mgmesytec/Vmmr16Parser.h
   mgmesytec/Sis3153Parser.h
   mgmesytec/Hit.h
-  mgmesytec/HitFile.h
   mgmesytec/SequoiaGeometry.h
   mgmesytec/BusGeometry.h
   )
@@ -80,7 +79,6 @@ set(MgConfigTest_INC
 )
 create_test_executable(MgConfigTest)
 
-
 set(MGHitTest_SRC
   mgmesytec/HitTest.cpp
   mgmesytec/Hit.cpp
@@ -88,18 +86,7 @@ set(MGHitTest_SRC
 set(MGHitTest_INC
   mgmesytec/Hit.h
   )
-create_test_executable(MGHitTest)
-
-
-set(MGHitFileTest_SRC
-  mgmesytec/HitFileTest.cpp
-  mgmesytec/Hit.cpp
-  )
-set(MGHitFileTest_INC
-  mgmesytec/HitFile.h
-  mgmesytec/Hit.h
-  )
-create_test_executable(MGHitFileTest SKIP_MEMGRIND)
+create_test_executable(MGHitTest SKIP_MEMGRIND)
 
 set(Sis3153ParserTest_SRC
   mgmesytec/Sis3153ParserTest.cpp
